

\name{dispersion}
\alias{dispersion}

\title{
Esimate both common and tagwise dispersion parameters for a dataset with a prescribed experimental design.
}
\description{
This function provides dispersion parameter estimates. The imputs are normalized RNA-seq read count data, an experimental design vector and a logical variable indicating whether the samples are matched or not. 
The output is a list containing the estimated common dispersion "commonDispersion" and the estimated tagwise dispersion vector "tagwiseDispersion".
}
\usage{
dispersion(M,condition,matched,method="edgeR")
}

\arguments{
  \item{M}{
  A list generated by the function "normalization", which contains the normalized read counts matrix 
  and the estimated size factor.
  }
  \item{condition}{
  A vector indicating the condition statuses of all samples.
  }
  \item{matched}{
  A logical variable indicating whether the samples from various conditions are matched (TRUE) or not (FALSE).
  }
  \item{method}{
  The method used to estimate the dispersion parameters. This should be either "edgeR" or "DESeq2". The default normalization method is "edgeR". 
  }
}
\value{

 \item{count}{Normalized read counts matrix, rows for genes and columns for samples.}

 \item{sizeFactor}{Estimated size factor.}

 \item{condition}{A vector indicating the condition statuses of all samples.}

 \item{commonDispersion}{Estimated common dispersion parameter.}
 
 \item{tagwiseDispersion}{Estimated tagwise dispersion parameter.}
 
 \item{matched}{A logical variable indicating whether the samples from various conditions are matched (TRUE) or not (FALSE).}

}
\references{
Kang G, Du L, Zhang H (2016). multiDE: A dimension reduced model based statistical method for differential expression analysis using RNA-sequencing data with multiple treatment conditions.
}
\author{
Guangliang Kang and Hong Zhang
} 

\examples{

data(psoriatic);

count = as.matrix(psoriatic[,-1]);

M = normalization(count,method="median");

condition = c(rep(1:8,each=3));

M = dispersion(M,condition=condition,matched=TRUE,method="edgeR");


}


\keyword{Dispersion estimation}



