\name{multiDE-package}
\alias{multiDE-package}
\docType{package}
\title{
Differential expression (DE) analysis for RNA-seq data with multiple treatment conditions
}
\description{
This package provides a statistical method for identifying DE genes using RNA-seq read count data from multiple conditions. Each condition has at least two samples, which can be either unmatched or matchced. The efficiency of this method is much improved over the existing methods through the use of a dimension reduced ANOVA model.
}
\details{
\tabular{ll}{
Package: \tab multiDE\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-04-17\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Guangliang Kang and Hong Zhang

Maintainer: Hong Zhang <zhangh@fudan.edu.cn>

}
\references{
Kang G., Du L., Zhang H. (2016). multiDE: A dimension reduced model based statistical method for differential expression analysis using RNA-sequencing data with multiple treatment conditions.
}

\keyword{ RNA-seq }  
