\name{multiDE}
\alias{multiDE}

\title{
Differential expression analysis for RNA-seq data with multiple treatment conditions
}
\description{
This function provides DE analysis results. The imput is a list generated by the function "dispersion", and the outputs include p-values, fold changes, and the variances of fold change for each gene. 
}
\usage{
multiDE(M,n.top=nrow(M$count))
}

\arguments{
  \item{M}{A list generated by the function "dispersion", which contains estimated dispersion parameters, condition statuses, and a logical variable indicating whether the samples are matched or not.
}
  \item{n.top}{The number of genes used for estimating parameter vector u, which should be smaller or equal to the number of all genes. The default value is the number of all genes.
}
}
\value{

 \item{log2FoldChange}{log2-fold changes.}

 \item{p.value}{P-values for DE analysis.}
 
 \item{u}{Estimation of parameters u.}
 
 \item{sd.log2FoldChange}{Standard errors of estimated log2-fold changes.}

}
\references{
Kang G, Du L, Zhang H (2016). multiDE: A dimension reduced model based statistical method for differential expression analysis using RNA-sequencing data with multiple treatment conditions.
}
\author{
Guangliang Kang and Hong Zhang
}
\examples{

data(cell);

count = as.matrix(cell[,-1]);

M = normalization(count,method="median");

M = dispersion(M,condition=rep(1:3,each=3),matched=FALSE,method="edgeR");

result = multiDE(M,n.top=1e3);

}


\keyword{Multiple conditions}

